Split graph retriever to query parser and graph retriever
